refactored chat component
